+++ proyect/src/main/java/app/domain/models/Transfer.java
@@ -1,7 +1,9 @@
 package app.domain.models;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
+import app.domain.models.enums.TransferStatus;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -14,7 +16,7 @@ public class Transfer{
     private long idTransfer;
     private BankAccount OriginAccount;
     private BankAccount DestinationAccount;
-    private double amount;
+    private BigDecimal amount;
     private Date creationDate;
     private Date approvalDate;
     private TransferStatus transferStatus;
